Improved the hit/mis tracking.
Only count things which are actually moved to the read buffer. Count
bytes instead of pages.

+++ seep-common/src/main/java/uk/ac/imperial/lsds/seep/core/DatasetMetadata.java
@@ -6,12 +6,12 @@ public class DatasetMetadata {
 	private long size;
 	private boolean inMem;
 	private long creationCost;
-	private int diskAccess;
-	private int memAccess;
+	private long diskAccess;
+	private long memAccess;
 	
 	public DatasetMetadata() { }
 	
-	public DatasetMetadata(int datasetId, long size, boolean inMem, long creationCost, int diskAccess, int memAccess) {
+	public DatasetMetadata(int datasetId, long size, boolean inMem, long creationCost, long diskAccess, long memAccess) {
 		this.datasetId = datasetId;
 		this.size = size;
 		this.inMem = inMem;
@@ -36,11 +36,11 @@ public class DatasetMetadata {
 		return creationCost;
 	}
 	
-	public int getDiskAccess() {
+	public long getDiskAccess() {
 		return diskAccess;
 	}
 	
-	public int getMemAccess() {
+	public long getMemAccess() {
 		return memAccess;
 	}
